% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.r
\name{checkIntegrity}
\alias{checkIntegrity}
\title{Checks integrity of sheets in the user input CSV file}
\usage{
checkIntegrity(dta.metab, dta.smetab, dta.sdata, dta.vmap, dta.models,
  dict_metabnames)
}
\arguments{
\item{dta.metab}{dta.metab}

\item{dta.smetab}{dta.smetab}

\item{dta.sdata}{dta.sdata}

\item{dta.vmap}{dta.vmap}

\item{dta.models}{dta.models}

\item{dict_metabnames}{dict_metabnames}
}
\description{
Checks integrity of sheets in the user input CSV file
}
\keyword{internal}
