% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaMP.R
\name{chemClassEnrichment}
\alias{chemClassEnrichment}
\title{Chemical Class Enrichment}
\usage{
chemClassEnrichment(df, metabdata, db.version = NULL, pvalue.adj = 0.05)
}
\arguments{
\item{df}{The "Effects" data frame returned from function \code{\link{runModel}}}

\item{metabdata}{Return object from \code{\link{readCOMETSinput}}}

\item{db.version}{The RaMP SQLite database version to use.
The default is the most recent version.}

\item{pvalue.adj}{The maximum BH-adjusted p-value for a metabolite
to be included in the analysis.}
}
\value{
A data frame
}
\description{
Chemical Class Enrichment
}
\details{
The set of all possible metbolite id types are:
LIPIDMAPS, pubchem, hmdb, chemspider, chebi, CAS, wikidata, swisslipids, 
kegg, lipidbank, plantfa, and kegg_glycan.
The \bold{METABOLITES} sheet in the input Excel file will need to have at least
one of the above id types as a column (case-insensitive). If the \bold{METABOLITES} sheet
contains more than one id type, then the metabolite ids passed into 
\code{\link[RaMP]{chemicalClassEnrichment}} depends on the order of the id type
columns. For example, if the \bold{METABOLITES} sheet contains columns HMDB (column F)
and pubchem (column C), then the metbolite ids will taken from the pubchem column and
any missing ids from the pubchem column will be assigned from the HMDB column.
}
\examples{
dir <- system.file("extdata", package="RcometsAnalytics", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInputAge.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
modeldata <- getModelData(exmetabdata,exposures="age", modelspec="Interactive",
	outcomes=c("lactose","lactate"))
obj <- runModel(modeldata,exmetabdata, cohortLabel="DPP")
# ret <- chemClassEnrichment(obj$Effects, exmetabdata, pvalue.adj=1)
}
