% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_clogit.R
\name{clogit.options}
\alias{clogit.options}
\title{options list for clogit}
\description{
A list of 2:
\itemize{
\item{\code{method}}{ One of: "exact", "approximate", "efron", "breslow".
            The default is "exact".}
\item{\code{weights}}{ A variable name to specify weights. The default is NULL.}
}
}
\examples{
model.options <- list(method="efron", weights="weightVarInData")
}
