% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_pcorr.R
\name{correlation.options}
\alias{correlation.options}
\title{options list for \code{model="correlation"}}
\description{
A list of 1:
\itemize{
\item{\code{method}}{ Correlation method to use. It must be one of 
                  "spearman", "pearson", "kendall".
                        The default value is "spearman".}
}
}
\examples{
model.options <- list(method="pearson")
}
