% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_coxph.R
\name{coxph.options}
\alias{coxph.options}
\title{options list for coxph}
\description{
A list of 6:
\itemize{
\item{\code{ties}}{ One of: "efron", "breslow", "exact".
            The default is "efron".}
\item{\code{robust}}{ TRUE or FALSE for computing a robust covariance matrix).
                    The default is FALSE.}
\item{\code{weights}}{ A variable name to specify weights. The default is NULL.}
\item{\code{singular.ok}}{ See \code{\link[survival]{coxph}}. The default is TRUE.}
\item{\code{Surv.type}}{ See the \code{type} option in \code{\link[survival]{Surv}}. The default is NULL.}
}
}
\examples{
model.options <- list(robust=TRUE, weights="weightVarInData")
}
