% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputColumns.R
\name{Effects}
\alias{Effects}
\title{Effects table}
\description{
The Effects data frame contains the estimates
 for each exposure, and  will contain multiple rows
 for categorical exposure variables.
Depending on the model run, 
all the below names may not appear in the data frame.
\itemize{
\item{\code{corr}}{ The correlation between \code{term} and the outcome}
\item{\code{estimate}}{ The regression coefficient of \code{term}}
\item{\code{exposurespec}}{ Exposure variable}
\item{\code{outcomespec}}{ Outcome variable}
\item{\code{p.value}}{ The p-value of the test}
\item{\code{run}}{ Run number that can be used to link with the
                   \code{\link{ModelSummary}} table}
\item{\code{statistic}}{ The test statistic for \code{term}}
\item{\code{std.error}}{ The standard error of \code{estimate}}
\item{\code{stratavar}}{ Stratum label or variable}
\item{\code{strata}}{ Stratum number or level}
\item{\code{term}}{ Exposure variable or dummy variable name}
}
}
\details{
Missing values will appear if a model did not converge, produced an error,
   or not run because of too many missing values in the outcome.
}
