% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_transform.R
\name{file.list}
\alias{file.list}
\title{list to describe or transform a file for meta-analysis}
\description{
A list of 20:
\itemize{
\item{\code{file}}{ The complete path to the file. Must be specified.}
\item{\code{cohort}}{ The name of the cohort. Use this option to set or change
         the name of the cohort.
         Must be specified for files not obtained from version 3.0 of RcometsAnalytics.}
\item{\code{sep}}{ The file delimiter. For example "," for csv files, "\\t" for
      tab-delimited files. The default is NULL.}
\item{\code{correlation}}{ 0 or 1 for correlation results. The default is 0, 
          so that the file is assumed to be from a non-correlation model.}
\item{\code{estimate.col}}{ Column name for the beta estimate or correlation. 
                        The default value is "estimate".}
\item{\code{se.col}}{ Column name for the standard error. Needed for non-correlation models.
                     The default value is "std.error".}
\item{\code{nobs.col}}{ Column name for the number of subjects. If such a column
       does not exist in the file, set \code{nobs} to a value.
                    The default is "nobs".}
\item{\code{nobs}}{ Number of observations. Must be specified if \code{nobs.col}
                    is not specified. The default is NULL.}
\item{\code{outcome.col}}{ Column name for the outcome. If such a column does not
                  exist in the file, then set \code{outcome.name}.
                           The default is "outcome_uid".}
\item{\code{outcome.name}}{ The name of the outcome variable. Must be specified if
                   \code{outcome.col} is not specified. No default.}
\item{\code{exposure.col}}{ Column name for the exposure. If such a column does not
                  exist in the file, then set \code{exposure.name}.
                           The default is "exposure_uid".}
\item{\code{exposure.name}}{ The name of the exposure variable. Must be specified if
                   \code{exposure.col} is not specified. No default.}
\item{\code{stratavar.col}}{ Column name for the stratification variable. 
                          The default is NULL.}
\item{\code{strata.col}}{ Column name for the stratification values. 
                          The default is NULL.}
\item{\code{strata.name}}{ The name of the stratification variable.
                 Use this option to change the name of the stratification variable or
                to add a stratification variable to the file. When adding a
                stratification variable, set \code{strata.value} also.
                The default is NULL.}
\item{\code{strata.value}}{ The constant value of the stratification.
                The default is NULL.}
\item{\code{model.col}}{ Column name for the model.  
                          The default is NULL.}
\item{\code{model.name}}{ The name of the model.
                 Use this option to change or add the model name to the results.
                The default is NULL.}
\item{\code{change.col.values}}{ A list of sublists, where each sublist has
         elements "col", "old", and "new". Use this option to change the 
         values of a (exposure or stratification) column to new values.
     For example, list(list(col="smoke", old=c("never", "current", "former"),
                            new=c(0, 1, 2))) will change the values
        "never" to 0, "current" to 1, and "former" to 2 in the "smoke" column.
                The default is NULL.}
\item{\code{where}}{Vector of strings with a variable name, 
       a comparison operator (e.g. "<", ">", "<=", ">=", "!=", "="), and a value.  
       The strings are combined by the logical AND (&) operator.
       For example, \code{where = c("study = A", "age > 50")} uses all subjects
       with age > 50 and in study A. This option would primarily be used if the file contained
       results from multiple cohorts and each cohort needed to be included
       in the meta-analysis. }
\item{\code{new.ref.value}}{ The new reference value for a categorical.
               exposure variable. This option will update the Effects data frame
               based on the new reference value. Note that the updated p-values
               will be based on a Wald test regardless of the original test.
               This option is applied before the option \code{change.col.values}.
               This option is only for results files from non-correlation
               models obtained using version 3.0 of RcometsAnalytics.
                The default is NULL.}
}
}
\details{
For any type of file, the parameter \code{file} must be specified. For files not obtained
        from version 3.0 of RcometsAnalytics, other parameters may need to be specified.
        This list can be used with the functions \code{\link{runMeta}} and
        \code{\link{meta_transform}}.
}
\examples{
# Suppose \code{f} is a file containing results from a linear regression model with
#  the metabolites as the outcomes in column called "metabolite", and the exposure
#  was a variable called "age" which is not a column in \code{f}. Also, 572 subjects
# were included in the analysis and the cohort name is "study_XYZ".
f <- "path_to_file"
file.list <- list(file=f, sep=",", cohort="study_XYZ", outcome.col="metabolite",
                       exposure.name="age", nobs=572)
}
