% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03b_getLM.r
\name{getLM}
\alias{getLM}
\title{Calculate lmelation matrix}
\usage{
getLM(modeldata, metabdata, cohort = "")
}
\arguments{
\item{modeldata}{list from function getModelData}

\item{metabdata}{metabolite data list}

\item{cohort}{cohort label (e.g DPP, NCI, Shanghai)}
}
\value{
data frame
}
\description{
Calculate lmelation matrix
}
