% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_glm.R
\name{glm.options}
\alias{glm.options}
\title{options list for glm}
\description{
A list of 6:
\itemize{
\item{\code{family}}{ One of: "binomial", "gaussian", "Gamma", "inverse.gaussian",
            "poisson", "quasi", "quasibinomial", "quasipoisson".
            The default is "gaussian".}
\item{\code{link}}{ NULL or a string for the link function to use (see \code{\link[stats]{family}}).
                    The default is to use the canonical link for \code{family}.}
\item{\code{weights}}{ A variable name to specify weights. The default is NULL.}
\item{\code{offset}}{ A variable name to specify an offset. The default is NULL.}
\item{\code{control}}{ See \code{\link[stats]{glm}}.
                     The default is \code{\link[stats]{glm.control}}.}
\item{\code{singular.ok}}{ See \code{\link[stats]{glm}}. The default is TRUE.}
}
}
\examples{
model.options <- list(family="binomial", weights="weightVarInData")
}
