% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel_lm.R
\name{lm.options}
\alias{lm.options}
\title{options list for lm}
\description{
A list of 4:
\itemize{
\item{\code{weights}}{ A variable name to specify weights. The default is NULL.}
\item{\code{offset}}{ A variable name to specify an offset. The default is NULL.}
\item{\code{tol}}{ See \code{\link[stats]{lm}}. The default is 1e-7.}
\item{\code{singular.ok}}{ See \code{\link[stats]{lm}}. The default is TRUE.}
}
}
\examples{
model.options <- list(weights="weightVarInData")
}
