% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaExt.R
\name{mergeResultsFiles}
\alias{mergeResultsFiles}
\title{This function will combine output files assumed to be from the same cohort and same model.}
\usage{
mergeResultsFiles(filevec, out.file, op = NULL)
}
\arguments{
\item{filevec}{Character vector of files that contain
the output files from \code{\link{runModel}} or \code{\link{runAllModels}}. 
Valid file extensions are ".xlsx", and ".rda".}

\item{out.file}{NULL or a file to save the results. The file extension must
be either ".xlsx", or ".rda".}

\item{op}{A list containing the options to use or NULL. The options are \code{precedence}
and \code{check.consistency} with default values of "nobs" and
TRUE respectively (see details).
The default is NULL.}
}
\value{
List of data frames containing the merged results
}
\description{
This function will combine output files assumed to be from the same cohort and same model.
}
\details{
This function will merge output files containing results from the same model
         and same cohort. Within a file, metabolites with missing results (
         \code{estimate} and \code{std.error}) will be excluded.
         The files can have different and overlapping metabolites.
         When there is an overlapping metabolite among the files, the option
         \code{precedence="nobs"} will choose the file with the largest number 
         of subjects for that metabolite. With \code{precedence="data"}, the first
         file in \code{filevec} that contains the metabolite will be chosen.
         The option \code{check.consistency} will check for the same 
         outcome, exposure, and stratification variable names across the files.
}
