% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta_calc}
\alias{meta_calc}
\title{Main function that performs the meta-analysis calculations.}
\usage{
meta_calc(beta, se)
}
\arguments{
\item{beta}{matrix or vector of betas. If a matrix, then the rows
represent the metabolites and the columns represent
the cohorts.}

\item{se}{matrix or vector of the standard errors for beta. This object
must be in the same order and have the same dimension as beta.}
}
\value{
List containing the results from fixed-effect and random-effect 
        meta-analyses along with Cochran's Q test for heterogeneity.
}
\description{
Main function that performs the meta-analysis calculations.
}
