% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_runAllMeta_options.R
\name{meta_opfile}
\alias{meta_opfile}
\title{Options file for meta-analyses}
\description{
An Excel file containing models and options for the \code{\link{runAllMeta}} function
}
\details{
The file should contain sheets \bold{META_MODELS} and \bold{META_TYPES}. 
Each sheet is optional. The \bold{META_MODELS} sheet should have a column
called \code{MODEL} containing the models for meta-analysis. 
This list of models can be a subset of models defined by the input files.
This sheet can also have a optional column called \code{MODEL_TYPE} that
links to the \bold{META_TYPES} sheet defining the options for each model.
See the example file in /inst/extdata/cometsMetaInput.xlsx.
}
