% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_transform.R
\name{meta_transform}
\alias{meta_transform}
\title{Function for transforming a set of results data}
\usage{
meta_transform(file.obj, out.file = NULL)
}
\arguments{
\item{file.obj}{list of type \code{\link{file.list}}}

\item{out.file}{NULL or the name of an output file to save the results.
The file extension must be ".xlsx" or ".rda".}
}
\value{
A list of objects with names \code{\link{ModelSummary}},
       \code{\link{Effects}}, \code{\link{Errors_Warnings}},
       \code{\link{Table1}}, \code{Info}. \cr
}
\description{
Function for transforming a set of results data
}
\details{
This function will transform or update a file of model results
         so that the results can be used for a meta-analysis. 
         The types of changes that can be made are:\cr
         1. Column values can be modified. For example, if the smoking status
            values 'never', 'current', and 'former' need to be changed to 0, 1, 2.\cr
         2. A column with a constant value can be added. For example, a column
            with the number of subjects used in the analysis can be added.  
            Or for instance, for a file from an all-female study, a stratification
            column containing the value 'female' can be added. \cr
         3. The file can be subsetted if it contains results from multiple models
            or multiple cohorts.\cr
         4. The results can be updated to reflect a different reference value
            of a categorical exposure variable.
         5. The name of the cohort can be renamed.\cr
         6. The name of the model can be renamed.
}
