% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputColumns.R
\name{ModelSummary}
\alias{ModelSummary}
\title{ModelSummary table}
\description{
The ModelSummary data frame contains one row of model
summary results for each exposure/outcome combination.
Depending on the model run, 
all the below names may not appear in the data frame.
\itemize{
\item{\code{adjspec}}{ Original adjustment variables specified}
\item{\code{adjvars}}{ Adjustment variables included in the model}
\item{\code{adjvars.removed}}{ Adjustment variables removed from the model}
\item{\code{adj_uid}}{ Adjustment variable universal ids}
\item{\code{adj.r.squared}}{ Adjusted R-squared}
\item{\code{aic}}{ Akaike information criterion}
\item{\code{bic}}{ Bayesian information criterion}
\item{\code{cohort}}{ String passed into \code{\link{runModel}}}
\item{\code{converged}}{ TRUE or FALSE for model convergence}
\item{\code{deviance}}{ Deviance of the fitted model}
\item{\code{df.null}}{ NULL model degrees of freedom}
\item{\code{df.residual}}{ Residual degrees of freedom}
\item{\code{exposure}}{ Exposure variable}
\item{\code{exposure_uid}}{ Exposure universal id}
\item{\code{exposurespec}}{ Exposure variable}
\item{\code{loglik}}{ Log-likelihood of the fitted model}
\item{\code{message}}{ Error message produced from the modeling function}
\item{\code{model}}{ Model label from \code{\link{getModelData}}}
\item{\code{model_function}}{ Model function used in \code{\link{runModel}}}
\item{\code{nobs}}{ Number of observations used}
\item{\code{null.deviance}}{ Deviance of the NULL model}
\item{\code{outcome}}{ Outcome variable}
\item{\code{outcomespec}}{ Outcome variable}
\item{\code{outcome_uid}}{ Outcome universal id}
\item{\code{run}}{ Run number that can be used to link with the 
                   \code{\link{Effects}} table}
\item{\code{runmode}}{ "Batch" or "Interactive"}
\item{\code{r.squared}}{ R-squared, the fraction of variance explained by the model}
\item{\code{sigma}}{ Square root of the estimated variance of the random error}
\item{\code{stratavar}}{ Stratum variable(s)}
\item{\code{strata}}{ Stratum level(s)}
\item{\code{term}}{ Variable in the model}
\item{\code{wald.pvalue}}{ P-value from the Wald test of the given variable. 
                          Note that this test may be a multi-df test if the
                          variable is categorical.}
}
}
\details{
Missing values will appear if a model did not converge, produced an error,
   or not run because of too many missing values in the outcome.
}
