% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{options}
\alias{options}
\title{options list}
\description{
A list of 12:
\itemize{
\item{\code{check.cor.cutoff}}{ Cutoff value to remove highly correlated columns in the
                        design matrix. The default value is 0.97.}
\item{\code{check.cor.method}}{ Correlation method to remove highly correlated columns 
                     in the design matrix. It must be one of 
                  "spearman", "pearson", "kendall".
                        The default value is "spearman".}
\item{\code{check.design}}{ TRUE or FALSE to check the design matrix for linearly
                    dependent columns, highly correlated columns, or for 
                    being ill-conditioned whenever it is updated.
                    The default is TRUE.}
\item{\code{check.illCond}}{ TRUE or FALSE to check for an ill-conditioned design matrix.
                     The default is TRUE.}
\item{\code{check.nsubjects}}{ Minimum number of subjects. The default is 25.}
\item{\code{max.nstrata}}{ The maximum number of strata for a stratified analysis.
                   The default is 10.}
\item{\code{model}}{ String for the model function. Currently, it must be one of
                     "correlation", "lm", "glm", "coxph", or "clogit". The default is "correlation".} 
\item{\code{model.options}}{ List of options specific for the model. 
       See \code{\link{correlation.options}}, \code{\link{glm.options}},
       \code{\link{lm.options}}, \code{\link{coxph.options}} and \code{\link{clogit.options}} for options specific to \cr
\code{model="correlation", "lm", "glm", "coxph", "clogit"} respectively.        
                     The default is NULL.} 
\item{\code{output.ci_alpha}}{ Confidence interval level for estimated from glm models. This
                        option must be a number >= 0 and < 1, where 0 is for not creating confidence intervals.
                        The default value is 0.95.}
\item{\code{output.Effects}}{ A string to define the terms output in the returned \code{\link{Effects}}
                           and \code{\link{ModelSummary}} data frames. Currently, it must be "exposure" or "all".
                          If set to "all", then summary statistics for the exposure
                           and adjustment variables will be output. Otherwise, only
                           summary statistics for the exposure will be output.
                           This option is ignored with \code{model = "correlation"}.
                           The default is "exposure".}
\item{\code{output.exp_parms}}{ TRUE, FALSE or NULL to exponentiate glm parameter estimates. 
                        Standard errors are obtained from the delta method. 
                        The default is NULL, so that estimates from glm models with 
                        family="binomial" will be exponentiated, and not otherwise.}
\item{\code{output.metab.cols}}{ Character vector of column names in the \code{METABOLITES}
                            sheet to be output in the \code{\link{ModelSummary}} and \code{\link{Effects}}
                            data frames. Metabolite ids are matched first using the 
                            \code{outcomespec} column and then using the \code{exposurespec} column.  
                        The default is "metabolite_name".}
\item{\code{output.ModelSummary}}{ A string to defines the columns output in the returned 
                           \code{\link{ModelSummary}} data frame. Currently, it must be "anova" or "all".
                           This option is ignored with \code{model = "correlation"}.
                           The default is "anova".}
}
}
\details{
Before any analysis is performed, an initial design matrix is
         constructed using the above options as follows: \cr
         1. Adjustment variables with less than two distinct non-missing
            values, or with less than \code{check.nsubjects} non-missing
            values are removed. \cr
         2. The design matrix is created from the remaining adjustment variables
            and any linearly dependent columns are removed. \cr
         3. If \code{check.cor.cutoff > 0}, then highly correlated columns
            are removed by computing the correlation matrix 
            \code{cor_matrix = cor(design_matrix, method=check.cor.method)}. \cr
         4. If \code{check.illCond} is TRUE, then the function
         \code{subselect::trim.matrix} is applied 
            to the correlation matrix to determine if additional
            columns will be removed. \cr \cr
If \code{check.design} is TRUE, then steps 2-4 are repeated each time an exposure
variable is added to the design matrix or when rows of the design matrix are removed
due to missing values in an outcome variable.
}
\examples{
# Logistic regression with all default options
model.op <- list(family="binomial")
op <- list(model="glm", model.options=model.op)
# Compute Pearson correlations requiring at least 100 subjects
model.op <- list(method="pearson")
op <- list(model="correlation", check.nsubjects=100, model.options=model.op)
}
