% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writingfunctions.r
\name{OutputListToExcel}
\alias{OutputListToExcel}
\title{Create an excel xlsx file from a list of data frames}
\usage{
OutputListToExcel(filename, obj)
}
\arguments{
\item{filename}{Name of file and can include path. It must have a ".xlsx" extension.}

\item{obj}{List of data frames or matrices}
}
\description{
Create an excel xlsx file from a list of data frames
}
