% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaMP.R
\name{pathwayEnrichment}
\alias{pathwayEnrichment}
\title{Pathway Enrichment}
\usage{
pathwayEnrichment(
  df,
  metabdata,
  db.version = NULL,
  pvalue.adj = 0.05,
  filter.p = 0.1,
  filter.p.type = "fdr",
  perc_analyte_overlap = 0.5,
  min_pathway_tocluster = 2,
  perc_pathway_overlap = 0.5
)
}
\arguments{
\item{df}{The "Effects" data frame returned from function \code{\link{runModel}}}

\item{metabdata}{Return object from \code{\link{readCOMETSinput}}}

\item{db.version}{The RaMP SQLite database version to use.
The default is the most recent version.}

\item{pvalue.adj}{The maximum adjusted p-value for a metabolite
to be included in the analysis.}

\item{filter.p}{See argument \code{pval_cutoff} in \code{RaMP::FilterFishersResults}.}

\item{filter.p.type}{See argument \code{pval_type} in \code{RaMP::FilterFishersResults}.}

\item{perc_analyte_overlap}{See argument \code{perc_analyte_overlap} in
\code{RaMP::findCluster}.}

\item{min_pathway_tocluster}{See argument \code{min_pathway_tocluster} in
\code{RaMP::findCluster}.}

\item{perc_pathway_overlap}{See argument \code{perc_pathway_overlap} in
\code{RaMP::findCluster}.}
}
\value{
A data frame
}
\description{
Pathway Enrichment
}
\details{
The \bold{RaMP} R package must be installed prior to calling
this function. An attempt to load RaMP will be done through the 
\code{\link[base]{requireNamespace}} function.
 
The set of all possible metbolite id types are:
LIPIDMAPS, pubchem, hmdb, chemspider, chebi, CAS, wikidata, swisslipids, 
kegg, lipidbank, plantfa, and kegg_glycan.
The \bold{METABOLITES} sheet in the input Excel file will need to have at least
one of the above id types as a column (case-insensitive). If the \bold{METABOLITES} sheet
contains more than one id type, then the metabolite ids passed into 
\code{\link[RaMP]{chemicalClassEnrichment}} depends on the order of the id type
columns. For example, if the \bold{METABOLITES} sheet contains columns HMDB (column F)
and pubchem (column C), then the metbolite ids will taken from the pubchem column and
any missing ids from the pubchem column will be assigned from the HMDB column.
}
\examples{
dir <- system.file("extdata", package="RcometsAnalytics", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInputAge.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
modeldata <- getModelData(exmetabdata,exposures="age", modelspec="Interactive",
	outcomes=c("lactose","lactate"))
obj <- runModel(modeldata,exmetabdata, cohortLabel="DPP")
# ret <- pathwayEnrichment(obj$Effects, exmetabdata, pvalue.adj=1)
}
