% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.r
\name{prdebug}
\alias{prdebug}
\title{debug by printing object with time time}
\usage{
prdebug(lab, x)
}
\arguments{
\item{lab}{label of object}

\item{x}{object}
}
\description{
debug by printing object with time time
}
\keyword{internal}
