% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RaMP.R
\name{ramp_options}
\alias{ramp_options}
\title{RaMP options list}
\description{
A list of 6:
\itemize{
\item{\code{dbname}}{ The default value is "ramp".}
\item{\code{username}}{ The default value is "root".}
\item{\code{conpass}}{ The default is "".}
\item{\code{host}}{ The default is "localhost".}
\item{\code{socket}}{ The default is "".}
\item{\code{pvalue.adj}}{ The maximum BH-adjusted p-value for a metabolite
                         to be included in the analysis.
                   The default is 0.05.}
}
}
\details{
See \code{\link[RaMP]{setConnectionToRaMP}} for details on the options
   \code{dbname}, \code{username}, \code{conpass}, \code{host},
   and \code{socket}.
}
