% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_readCOMETSinput.R
\name{readCOMETSinput}
\alias{readCOMETSinput}
\title{Read in Excel file that contains metabolite data, covariate data, 
models, and model options.}
\usage{
readCOMETSinput(file)
}
\arguments{
\item{file}{path of Excel file to be read in. This file must contain sheets
with names \bold{Metabolites}, \bold{SubjectMetabolites}, \bold{SubjectData}, \bold{VarMap}, 
and optionally \bold{Models}, \bold{ModelOptions} (see details).}
}
\value{
a list comprising of data and information needed for \code{\link{getModelData}}.
}
\description{
Read in Excel file that contains metabolite data, covariate data, 
models, and model options.
}
\details{
Additional information regarding each sheet in the input Excel file is given below. 

\bold{Metabolites} \cr
A table with the columns \code{METABID}, \code{METABOLITE_NAME}, and possibly other columns \cr
 of information about the metabolites. The \code{METABID} column is used
 for harmonizing the metabolite names across different cohorts when meta-analyses are performed. \cr
\bold{SubjectMetabolites} \cr
A table with the subject ids in the first column and metabolites as the other columns. \cr
\bold{SubjectData} \cr
A table with the subject ids in the first column and covariates as the other columns. \cr
\bold{VarMap} \cr
A table with at least the required columns \code{VARREFERENCE}, \code{COHORTVARIABLE}, 
\code{VARTYPE}, and \code{VARDEFINITION}. The \code{COHORTVARIABLE} column must contain names that match the
column names in the \bold{SubjectData} table. These names will be renamed to their
corresponding name in the \code{VARREFERENCE} column. The \code{VARTYPE} column 
should have values \code{continuous} or \code{categorical} for each row. \cr
\bold{Models} \cr
A table where each row represents a model to be run, and with columns \code{MODEL}, 
\code{OUTCOMES}, \code{EXPOSURE}, \code{ADJUSTMENT},
 \code{STRATIFICATION}, \code{WHERE}, and optionally \code{MODELSPEC}. All variable names in this
table must match variable names in the \code{VARREFERENCE} column of the \bold{VarMap} sheet.
The \code{MODEL} column is a label for the model. The \code{OUTCOMES} and \code{EXPOSURE} columns define the 
outcome and exposure variables for the model. Use \code{All metabolites} to specify
that all metabolite variables are to be included as outcomes or exposures, otherwise 
use a space separated list of variable names. The \code{ADJUSTMENT} column contains a
space separated list of covariate adjustment variables; use an empty cell for no covariate adjustment.
The \code{STRATIFICATION} column is used for stratified analyses, with a space separated list
of stratification variables. If more than one stratification variable is specified, then the strata
are defined by all unique combinations of the stratification variables that appear in the data.
The \code{WHERE} column is used to define a subset of subjects to include in the analysis,
and has the form \code{variable operator value}, where \code{operator} can be one of the
following \code{>, <, >=, <= !=, =}.
An example \code{WHERE} condition is \code{age > 50}, which will include all subjects older
than 50 in the analysis. Multiple \code{WHERE} conditions must be separated by a \code{&}. 
For example, \code{age > 50 & bmi >= 22} will include the subjects older than 50 AND with
bmi >= 22. Values in the \code{MODELSPEC} column must match with the \code{MODELSPEC} column
in the \bold{ModelOptions} sheet. 
This sheet is not required when running in interactive mode, but is required when
running in batch mode. \cr
\bold{ModelOptions} \cr
A table where each row specifies an option and has columns \code{MODELSPEC}, \code{FUNCTION},
\code{OPTION}, and \code{VALUE}. For an example sheet and additional information about this
sheet, see the Excel file \code{/extdata/cometsInput.xlsx}.
This sheet is optional, but is required when the \bold{Models} sheet contains the 
column \code{MODELSPEC}.
}
\examples{
dir <- system.file("extdata", package="RcometsAnalytics", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInputAge.xlsx")
exmetabdata <- readCOMETSinput(csvfile)

}
