% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_getCorr.r
\name{runCorr}
\alias{runCorr}
\title{Calculate correlation matrix for input model.}
\usage{
runCorr(modeldata, metabdata, cohort = "", op = NULL)
}
\arguments{
\item{modeldata}{list from function \code{\link{getModelData}}}

\item{metabdata}{metabolite data list from \code{\link{readCOMETSinput}}}

\item{cohort}{cohort label (e.g DPP, NCI, Shanghai)}

\item{op}{list of options (see \code{\link{options}}).}
}
\value{
A list of objects with names \code{\link{ModelSummary}},
       \code{\link{Effects}}, and \code{\link{Errors_Warnings}}. 
Attribute of this list includes ptime for processing time of model run.
}
\description{
Calculate correlation matrix for input model.
}
\details{
This function is a special case of \code{\link{runModel}}
         with the option \code{op$model = "correlation"}
}
\examples{
dir <- system.file("extdata", package="COMETS", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInputAge.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
modeldata <- getModelData(exmetabdata,exposures="age",modlabel="1 Gender adjusted",
	outcomes=c("lactose","lactate"), modelspec="Interactive")
corrmatrix <- runCorr(modeldata,exmetabdata, "DPP")
}
