% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_readCOMETSinput.R
\name{runDescrip}
\alias{runDescrip}
\title{This function provides a description of the input data (for categorical data, the number of samples of each type; for continous data, the median and other statistics for each variable)}
\usage{
runDescrip(readData)
}
\arguments{
\item{readData}{list from readComets}
}
\value{
a list with 2 data frames, continuous and categorical summaries. Type of variable is defined in varmap
}
\description{
This function provides a description of the input data (for categorical data, the number of samples of each type; for continous data, the median and other statistics for each variable)
}
\examples{
\dontrun{
dir <- system.file("extdata", package="COMETS", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInputAge.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
allmodeloutput <- runAllModels(exmetabdata)
# Get descriptive data
descdata <-runDescrip(exmetabdata)
OutputXLSResults(filename="corr",datal=descdata,cohort="DPP")
}
}
