% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{runMeta}
\alias{runMeta}
\title{This function allows users to run a single meta-analysis}
\usage{
runMeta(filevec, op = NULL)
}
\arguments{
\item{filevec}{Character vector of files that contain
the output files from \code{\link{runModel}} or \code{\link{runAllModels}}. 
Valid file extensions are ".xlsx", and ".rda".}

\item{op}{A list containing the options to use.
See \code{\link{meta_options}}.
The default is NULL.}
}
\value{
List of data frames containing the results and information
}
\description{
This function allows users to run a single meta-analysis
}
