% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModel.R
\name{runModel}
\alias{runModel}
\title{Main function for running the models}
\usage{
runModel(modeldata, metabdata, cohort = "", op = NULL)
}
\arguments{
\item{modeldata}{list from function \code{\link{getModelData}}}

\item{metabdata}{metabolite data list from \code{\link{readCOMETSinput}}}

\item{cohort}{cohort label (e.g DPP, NCI, Shanghai)}

\item{op}{list of options (see \code{\link{options}}).}
}
\value{
A list of objects with names \code{\link{ModelSummary}},
       \code{\link{Effects}}, and \code{\link{Warnings}}. 
Attribute of this list includes ptime for processing time of model run.
}
\description{
Main function for running the models
}
\details{
This function will check for stratification variables, and if
         present run within each stratum. The design matrix is
         checked for validity (see \code{\link{options}}).
}
\examples{
dir <- system.file("extdata", package="COMETS", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInputAge.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
modeldata <- getModelData(exmetabdata,colvars="age",modlabel="1 Gender adjusted",
	rowvars=c("lactose","lactate"))
obj <- runModel(modeldata,exmetabdata, "DPP")
}
