% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{showHClust}
\alias{showHClust}
\title{Show interactive heatmap using heatmaply_cor with hierarchical clustering}
\usage{
showHClust(
  ccorrList,
  strata = NULL,
  clust = TRUE,
  colscale = "RdYlBu",
  showticklabels = TRUE
)
}
\arguments{
\item{ccorrList}{correlation object (output of \code{\link{runCorr}})}

\item{strata}{Only valid if ccorrList is from a stratified analysis. 
If NULL, then results from the first stratum will be used in the plot.}

\item{clust}{TRUE or FALSE to show hierarchical clustering. The default is TRUE.}

\item{colscale}{colorscale, can be custom or named ("Hots","Greens","Blues","Greys","Purples") see \code{\link[heatmaply]{RColorBrewer_colors}}}

\item{showticklabels}{TRUE or FALSE to show axis labels. The default is TRUE.}
}
\value{
a heatmap with outcomes as rows and exposures in columns.
}
\description{
This function outputs a heatmap with hierarchical clustering.  It thus requires you to have at least 2 outcome and 2 exposure variables in your models.
}
\examples{
dir <- system.file("extdata", package="RcometsAnalytics", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInputAge.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
modeldata <- getModelData(exmetabdata, modelspec="Interactive",
exposures=c("age","bmi_grp"))
corrmatrix <-runCorr(modeldata,exmetabdata,"DPP")
showHClust(corrmatrix)
}
\references{
For colorscale reference: \code{\link[heatmaply]{RColorBrewer_colors}}
}
