% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{showModel}
\alias{showModel}
\title{Function that displays the first N rows of each data frame 
 in the the \code{\link{runModel}} output.}
\usage{
showModel(obj, nlines = 10)
}
\arguments{
\item{obj}{returned object from \code{\link{runModel}}}

\item{nlines}{number of lines to display (default 10)}
}
\description{
Function that displays the first N rows of each data frame 
 in the the \code{\link{runModel}} output.
}
\examples{
dir         <- system.file("extdata", package="COMETS", mustWork=TRUE)
csvfile     <- file.path(dir, "cometsInputAge.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
modeldata   <- getModelData(exmetabdata,modlabel="1 Gender adjusted")
result      <- runModel(modeldata,exmetabdata, cohortLabel="DPP")
showModel(result)
}
