% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFunctions.r
\name{trim.matrix}
\alias{trim.matrix}
\title{This function removes variables that cause the input design matrix to be ill-conditionned. The function was copied direclty from the package as a quickfix: https://cran.r-project.org/src/contrib/Archive/subselect/subselect_0.15.5.tar.gz}
\usage{
trim.matrix(mat, tolval = 10 * .Machine$double.eps)
}
\arguments{
\item{mat}{correlation matrix}

\item{totalval}{totalval}
}
\description{
This function removes variables that cause the input design matrix to be ill-conditionned. The function was copied direclty from the package as a quickfix: https://cran.r-project.org/src/contrib/Archive/subselect/subselect_0.15.5.tar.gz
}
\keyword{internal}
