% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputColumns.R
\name{Warnings}
\alias{Warnings}
\title{Warnings table columns}
\description{
Columns in the Warnings matrix. 
\itemize{
\item{\code{type}}{ WARNING or ERROR}
\item{\code{object}}{ The object that produced the warning or error. This is
                      typically a variable or a particular stratum.}
\item{\code{message}}{ Message describing the warning or error}
}
}
\details{
The kinds of warnings and errors stored in this matrix are ones that
   apply to all models or all outcomes for an exposure variable. An error
   message for a particular exposure-outcome pair will be stored in the
   \code{message} column of the \code{\link{ModelSummary}} table.
}
