% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputColumns.R
\name{Effects}
\alias{Effects}
\title{Effects table}
\description{
The Effects data frame contains the estimates
 for each exposure, and  will contain multiple rows
 for categorical exposure variables.
Depending on the model run and options specified, 
all the below names may not appear in the data frame.
\itemize{
\item{\code{corr}}{ The correlation between \code{term} and the outcome}
\item{\code{estimate}}{ The regression coefficient of \code{term}}
\item{\code{estimate.lower}}{ The lower confidence limit for \code{term}}
\item{\code{estimate.upper}}{ The upper confidence limit for \code{term}}
\item{\code{exp.estimate}}{ The exponentiated regression coefficient of \code{term}}
\item{\code{exp.estimate.lower}}{ The exponentiated lower confidence limit for \code{term}}
\item{\code{exp.estimate.upper}}{ The exponentiated upper confidence limit for \code{term}}
\item{\code{exp.std.error}}{ The standard error of \code{exp.estimate} from the delta method}
\item{\code{exposurespec}}{ Exposure variable}
\item{\code{model}}{ Model label from \code{\link{getModelData}}}
\item{\code{model_number}}{ Model number used in \code{\link{runAllModels}}}
\item{\code{outcomespec}}{ Outcome variable}
\item{\code{pvalue}}{ The p-value of the test}
\item{\code{pvalue.adj}}{ Adjusted p-values for exposure variables only}
\item{\code{run}}{ Run number that can be used to link with the
                   \code{\link{ModelSummary}} table}
\item{\code{statistic}}{ The test statistic for \code{term}}
\item{\code{std.error}}{ The standard error of \code{estimate}}
\item{\code{stratavar}}{ Stratum variable(s)}
\item{\code{strata}}{ Stratum level(s)}
\item{\code{term}}{ Variable in the model}
}
}
\details{
Missing values will appear if a model did not converge, produced an error,
   or not run because of too many missing values in the outcome. 
   Adjusted p-values are only computed for models with a single outcome variable
   or single exposure variable.
}
