% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_getModelData.R
\name{getModelData}
\alias{getModelData}
\title{Prepares data for the models to be run as specified in the input.  Can be run in interactive or batch mode.  Each model is checked for validity (correlation between predictors, zero variance, etc.).}
\usage{
getModelData(
  readData,
  modelspec = "Batch",
  modlabel = "",
  outcomes = "All metabolites",
  exposures = "",
  adjvars = NULL,
  strvars = NULL,
  wgtvar = NULL,
  offvar = NULL,
  timevar = NULL,
  groupvar = NULL,
  where = NULL
)
}
\arguments{
\item{readData}{List from \code{\link{readCOMETSinput}}}

\item{modelspec}{How model is specified (Interactive or Batch). The default is Batch}

\item{modlabel}{If batch, chosen model specified by batch mode (the MODEL column in
the Models sheet). If interactive, then the model label.}

\item{outcomes}{If Interactive, a vector of outcome variables (see \code{details}), the default is All metabolites)}

\item{exposures}{If Interactive, a vector of exposure variables (see \code{details})}

\item{adjvars}{If Interactive, a vector adjustment covariates (see \code{details})}

\item{strvars}{If Interactive, stratification covariates (see \code{details})}

\item{wgtvar}{If Interactive, a variable of weights (see \code{details})}

\item{offvar}{If Interactive, an offset variable (see \code{details})}

\item{timevar}{If Interactive, time variable(s) for survival models (see \code{details})}

\item{groupvar}{If Interactive, a group variable for conditional logistic models (see \code{details})}

\item{where}{users can specify which subjects to perform the analysis on by specifying this parameter. 
   'where' expects a vector of strings with a variable name, 
   a comparison operator (e.g. "<", ">", "<=", ">=", "!=", "="), and a value.  
   For example, \code{where = c("age>50","bmi > 22")} uses all subjects
   with age > 50 AND bmi > 22.  
Note that when running in Batch mode, rules in the \code{WHERE} column
of the \code{Models} sheet must be separated by a comma.}
}
\value{
a list comprising: \cr
1: subset data: gdta \cr
2: exposure variables: ccovs \cr
3: outcome variables: rcovs \cr
4: adjustment variables: acovs \cr
5: stratification variable: scovs \cr
6: model specification: modspec \cr
7: model label: modlab \cr
8: whether all metabolites vs all metabolites is run: allvsall \cr
9: weight variables: wgtcov \cr
10: offset variables: offcov
}
\description{
Prepares data for the models to be run as specified in the input.  Can be run in interactive or batch mode.  Each model is checked for validity (correlation between predictors, zero variance, etc.).
}
\details{
All metabolite variables specified should be listed in the \code{metabolite_name}
  column of the \code{Metabolites} sheet of the Excel file. All non-metabolite
  variables should be listed in the \code{VARREFERENCE} column of the
  \code{VarMap} sheet. The \code{wgtvar}, \code{offvar}, and \code{timevar} are only used
  for specific models. See the
  \code{model} option in \code{\link{options}}.
}
\examples{
dir <- system.file("extdata", package="RcometsAnalytics", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInputAge.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
modeldata <- getModelData(exmetabdata,modlabel="1 Gender adjusted")

}
