% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_runAllMeta_options.R
\name{meta_opfile}
\alias{meta_opfile}
\title{Options file for meta-analyses}
\description{
An Excel file containing models and options for the \code{\link{runAllMeta}} function
}
\details{
The file can contain sheets \bold{META_MODELS} and \bold{META_TYPES}. 
Each sheet is optional. The \bold{META_MODELS} sheet should have a column
called \code{MODEL} containing the models that the user wants to run 
meta-analyses for. 
If the \bold{MODELS} sheet is not given then meta-analyses for all models
 will be run.
This sheet can also have an optional column called \code{META_TYPE} that
links to the \code{META_TYPE} column in the \bold{META_TYPES} sheet to define
the options for each meta-analysis model.
The \bold{META_TYPES} sheet should contain columns \code{META_TYPE},  \code{OPTION},
and \code{VALUE}. 
See the example file in /inst/extdata/cometsMetaInput.xlsx.
If the \bold{META_TYPES} sheet is not specified, then the default values for all
 meta-analysis options will be used.
}
