% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_runAllMeta.R
\name{runAllMeta}
\alias{runAllMeta}
\title{This function allows users to run meta-analyses on all models or a subset of models}
\usage{
runAllMeta(filesFolders, out.dir, opfile = NULL)
}
\arguments{
\item{filesFolders}{Character vector of files and/or folders that contain
the output files from \code{\link{runAllModels}}. Valid file
extensions are ".xlsx", ".rda", ".zip", ".tar", and ".tar.gz". 
Zip and tar files must contain files with extensions
".xlsx" or ".rda".}

\item{out.dir}{Output directory to save the results for each model.}

\item{opfile}{Excel file containing the models and options.
See \code{\link{meta_opfile}}.
The default is NULL.}
}
\description{
This function allows users to run meta-analyses on all models or a subset of models
}
