% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_getCorr.r
\name{runCorr}
\alias{runCorr}
\title{Calculate correlations for input model.}
\usage{
runCorr(modeldata, metabdata, cohort = "")
}
\arguments{
\item{modeldata}{list from function \code{\link{getModelData}}}

\item{metabdata}{metabolite data list from \code{\link{readCOMETSinput}}}

\item{cohort}{cohort label (e.g DPP, NCI, Shanghai)}
}
\value{
data frame with each row representing the correlation for each combination of outcomes and 
exposures represented as specified in the model (*spec), label (*lab), and universal id (*_uid)
with additional columns for n, pvalue, method of model specification (Interactive or Batch), 
universal id for outcomes (outcome_uid) and exposures (exposure_uid)
name of the cohort, adjustment (adjvars) and stratification (stratavar,strata)  variables. 
Attribute of dataframe includes ptime for processing time of model run.
}
\description{
Calculate correlations for input model.
}
\details{
This function is a special case of \code{\link{runModel}}
         with the option \code{op$model = "correlation"}, 
         however for backwards compatibility, it returns a data frame as
         in the original version of the \bold{COMETS} R package.
}
\examples{
dir <- system.file("extdata", package="RcometsAnalytics", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInputAge.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
modeldata <- getModelData(exmetabdata,exposures="age",modlabel="1 Age",
	outcomes=c("lactose","lactate"), modelspec="Interactive")
corrmatrix <- runCorr(modeldata,exmetabdata, "DPP")
}
