% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_runModel.R
\name{runModel}
\alias{runModel}
\title{Main function for running the models}
\usage{
runModel(modeldata, metabdata, cohortLabel = "", op = NULL)
}
\arguments{
\item{modeldata}{list from function \code{\link{getModelData}}}

\item{metabdata}{metabolite data list from \code{\link{readCOMETSinput}}}

\item{cohortLabel}{cohort label (e.g DPP, NCI, Shanghai)}

\item{op}{list of options when running in \code{Interactive} mode (see \code{\link{options}}).}
}
\value{
A list of objects with names \code{\link{ModelSummary}},
       \code{\link{Effects}}, and \code{\link{Errors_Warnings}}. \cr
\bold{Important: check the \code{adjvars.removed} column in the
 \code{\link{ModelSummary}} data frame to see if any adjustment variables
 were dropped from the model, and use caution interpreting results
 when variables are removed. The \code{\link{Errors_Warnings}} object may
 also contain additional variables removed.} 
Attribute of this list includes ptime for processing time of model run.
}
\description{
Main function for running the models
}
\details{
This function will check for stratification variables, and if
         present run within each stratum. The design matrix is
         checked for validity (see \code{\link{options}}). 
         When running in \code{Batch} mode, the options are obtained from
         the \code{Options} sheet in the excel file. \cr
}
\examples{
dir <- system.file("extdata", package="RcometsAnalytics", mustWork=TRUE)
csvfile <- file.path(dir, "cometsInputAge.xlsx")
exmetabdata <- readCOMETSinput(csvfile)
modeldata <- getModelData(exmetabdata,exposures="age",modlabel="1 Gender adjusted",
	outcomes=c("lactose","lactate"))
obj <- runModel(modeldata,exmetabdata, cohortLabel="DPP")
}
